Train Scheduling Update Part 1
Updated the CTC UI to serve as the main landing page, added new buttons and reorganized the layout to accomadate the changes (since the CTC will now be used to access the separate sections of the simulation). Scheduled trains should now appear in a separate window.  TODO: Create functions for both delete and dispatching the trains and their appropriate behaviors. Test to make sure schedules are created