Documentation Updates, Additional Train Partial, CTC Updates
The CTC now displays the number of active trains. I updated the comments for the last edit dates that I forgot to include on some of my earlier pushes. The schedules now display different trains, but an error is created in the TrainModelUI with the list of Trains there being null (thought the track model displays a second one). I believe the issue with scheduling past 100 is that the current method can not handle manuevering around a switch. Will work on this tomorrow